Use structured data to pass around the weather info
The is the second part in order to display rich weather UI such as icons etc. Previously we are using a formated text for the list view display. This PR defined a new class DailyWeatherInfo to represent a daily weather info with fields such as high temp, low temp, weather icon ID, weather description and so on. So that we can render the weather UI flexiblely in the future.