Allow user input location with auto-completion support
This change adds a location input section on the Preferences page, which allows user to type in any city (with optional state/country) to query the weather data. The component is implemented by using AutoCompleteTextView with auto-complete support for the top 1000 US cities, and the user can enter any other location manually if wanted. The network utils has been updated accordingly to take the user input.