support GPS location to fetch weather data by lat/lng
Previously we use the user entered location (with auto-complete) on the Settings page to load the weather data. The PR add a second mode which accesses the GPS location to fetch the weather data if the user doesn't enter the location on the Settings page. If the GPS location is not available temparally, then displays an error message suggesting the user to enter the location manually on the Settings page to continue using the app.