Start a new Activity for weather rating
Would like to build a new feature "How do like the weather?" to encourge user to rate the weather and see how other people nearby like the weather. The PR implements the UI part to start a new activity. The next step is to use firebase to submit rating data and pull other people's aggregated rating based on the location.