set up firebase and support weather sentiment submission
This PR is the first part to build the location sentiment feature, which allow each user to rate the weather for a location. The data is stored in firebase in the nested structure of city->userid>ratingData. This storage structure ensures that each user will have at most one rating per city (can be updated). The ratingData includes the info of the rating, the distance between the user and the city center, and the timestamp when the rating happens. The distance and the timestemp will be used to calculate the aggregated weighted rating for each city later. i.e., (a) make sentiment from users who are nearby the location weight more in the average than users who are further away from the location and (b) make older sentiments count less. The aggregated weighted city rating will be built in a separate PR.