calculate the weighted rating from firebase data
The change added the firebase retrieval logic to fetch the existing weather rating data for a given location, and then use a Squash Half scoring function to compute the weight decay for each of the weather rating age and the weather rating distance from the location center. The age weight decays to half if it is one-week old and the distance weight decays to half if it is 10KM away. The decay curve is further controled by a power parameter. If a user choose to submit a rating, then the weighted rating will be updated immediately by using the firebase event listener. The change has some tweaks on the xml to make the UI look nicer.